Prevent shutdown/reboot/logoff if video encoding is in progress

+++ system/utilities/system.fish
@@ -13,24 +13,37 @@ switch $option
     case Lock
         swaylock --image (wallpaper --current)
     case Logout
+        if pgrep --exact ffmpeg &>/dev/null
+            notify-send --app-name system --icon dialog-warning --urgency critical 'LOGOUT ABORTED' 'A ffmpeg process is currently running.'
+            pw-play /usr/share/sounds/freedesktop/stereo/dialog-warning.oga
+            exit 1
+        end
         swaymsg exit
     case Sleep
         systemctl suspend
     case Hibernate
         set mounted_media (findmnt --output TARGET --noheadings --raw | string match '/media*')
         if test -n "$mounted_media"
-            notify-send --app-name system --icon dialog-warning --urgency critical \
-                "CANNOT HIBERNATE" \
-                "There is media mounted at\n$mounted_media"
+            notify-send --app-name system --icon dialog-warning --urgency critical 'HIBERNATE ABORTED' "There is media mounted at\n$mounted_media"
             pw-play /usr/share/sounds/freedesktop/stereo/dialog-warning.oga
             exit 1
         end
         wpctl set-volume @DEFAULT_AUDIO_SINK@ 50%
         systemctl hibernate
     case Reboot
+        if pgrep --exact ffmpeg &>/dev/null
+            notify-send --app-name system --icon dialog-warning --urgency critical 'REBOOT ABORTED' 'A ffmpeg process is currently running.'
+            pw-play /usr/share/sounds/freedesktop/stereo/dialog-warning.oga
+            exit 1
+        end
         wpctl set-volume @DEFAULT_AUDIO_SINK@ 50%
         systemctl reboot
     case Poweroff
+        if pgrep --exact ffmpeg &>/dev/null
+            notify-send --app-name system --icon dialog-warning --urgency critical 'POWEROFF ABORTED' 'A ffmpeg process is currently running.'
+            pw-play /usr/share/sounds/freedesktop/stereo/dialog-warning.oga
+            exit 1
+        end
         wpctl set-volume @DEFAULT_AUDIO_SINK@ 50%
         systemctl poweroff
     case '*'
